authorization using jwt

+++ src/items/items.service.ts
@@ -1,5 +1,5 @@
 import {
-  Inject,
+  BadRequestException,
   Injectable,
   InternalServerErrorException,
   NotFoundException,
@@ -9,11 +9,13 @@ import { ItemStatus } from './item-status.enum';
 import { UpdateItemDto } from './dto/update-item.dto';
 import { Item } from '../entities/item.entity';
 import { DeleteResult, Repository } from 'typeorm';
+import { InjectRepository } from '@nestjs/typeorm';
+import { User } from '../entities/user.entity';
 
 @Injectable()
 export class ItemsService {
   constructor(
-    @Inject('ITEM_REPOSITORY')
+    @InjectRepository(Item)
     private itemRepository: Repository<Item>,
   ) {}
 
@@ -34,7 +36,7 @@ export class ItemsService {
     return item;
   }
 
-  async create(createItemDto: CreateItemDto): Promise<Item> {
+  async create(createItemDto: CreateItemDto, user: User): Promise<Item> {
     const { name, price, description } = createItemDto;
     try {
       const saveItem = await this.itemRepository.create({
@@ -44,7 +46,8 @@ export class ItemsService {
         status: ItemStatus.ON_SALE,
         createdAt: new Date(),
         updatedAt: new Date(),
-      } as CreateItemDto);
+        user,
+      } as unknown as Item);
 
       return await this.itemRepository.save(saveItem);
     } catch (e) {
@@ -52,8 +55,13 @@ export class ItemsService {
     }
   }
 
-  async updateStatus(id: string): Promise<Item> {
+  async updateStatus(id: string, user: User): Promise<Item> {
     const item = await this.itemRepository.findOneBy({ id });
+    if (item.userId === user.id) {
+      throw new BadRequestException(
+        '自身が出品した商品を購入することは出来ません。',
+      );
+    }
     const updatedItem = {
       ...item,
       status: ItemStatus.SOLD_OUT,
@@ -65,8 +73,17 @@ export class ItemsService {
     return updatedItem;
   }
 
-  async update(id: string, updateItemDto: UpdateItemDto): Promise<Item> {
+  async update(
+    id: string,
+    updateItemDto: UpdateItemDto,
+    user: User,
+  ): Promise<Item> {
     const item = await this.itemRepository.findOneBy({ id });
+    if (item.userId !== user.id) {
+      throw new BadRequestException(
+        '他のユーザーが出品した商品の変更はできません。',
+      );
+    }
     const { name, price, description } = updateItemDto;
     const updatedItem = {
       ...item,
@@ -81,7 +98,13 @@ export class ItemsService {
     return updatedItem;
   }
 
-  async delete(id: string): Promise<DeleteResult> {
+  async delete(id: string, user: User): Promise<DeleteResult> {
+    const item = await this.findById(id);
+    if (item.userId !== user.id) {
+      throw new BadRequestException(
+        '他のユーザーが出品した商品は削除できません。',
+      );
+    }
     const response = await this.itemRepository.delete(id).catch((e) => {
       throw new InternalServerErrorException(e.message);
     });
